+++ src/main/java/com/sparta/quizdemo/order/service/OrderService.java
@@ -17,6 +17,7 @@ import com.sparta.quizdemo.order.repository.OrderItemRepository;
 import com.sparta.quizdemo.order.repository.OrderRepository;
 import com.sparta.quizdemo.common.entity.UserRoleEnum;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.json.JsonParseException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -36,6 +37,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class OrderService {
@@ -79,6 +81,7 @@ public class OrderService {
         lat2 = Double.parseDouble(xyMap.get("y")); // 위도
         lon2 = Double.parseDouble(xyMap.get("x")); // 경도
         city = xyMap.get("z").substring(1);
+        log.info(xyMap.get("z"));
 
         switch (city) {
             case "서울": distance = distance(37.330689, 126.593066, lat2, lon2);
@@ -111,12 +114,17 @@ public class OrderService {
 
         if (orderRequestDto.getPayment().equals(totalPrice)) {
             // 현재 유저의 order 생성
-            Order order = new Order(user, totalPrice, completeTime, orderRequestDto.getRequest());
+            Order order = new Order(user, totalPrice, completeTime, orderRequestDto.getRequest(), orderRequestDto.getOrderComplete());
             orderRepository.save(order);
+
+            List<Order> userOrderList = orderRepository.findAllByUserIdOrderByCreatedAtAsc(user.getId());
+            if (userOrderList.size() > 10) {
+                orderRepository.delete(userOrderList.remove(0));
+            }
 
             if (cartItemList != null) {
                 for (CartItem cartItem : cartItemList) {
-                    OrderItem orderItem = new OrderItem(cartItem, order);
+                    OrderItem orderItem = new OrderItem(cartItem, order, cartItem.getOptionList());
                     orderItemRepository.save(orderItem);
                     cartItemRepository.delete(cartItem);
                 }
@@ -128,22 +136,40 @@ public class OrderService {
         }
     }
 
-    public ResponseEntity<List<OrderResponseDto>> getOrderList() {
+    public ResponseEntity<List<OrderResponseDto>> getDoneOrderList() {
+        List<Order> orderList = orderRepository.findAllByOrderByCreatedAtDesc();
+        if (orderList.isEmpty()) {
+            throw new NullPointerException("주문이 없습니다.");
+        } else {
+            List<OrderResponseDto> orderResponseDtoList = new ArrayList<>();
+            for (Order order : orderList) {
+                if (order.getOrderComplete()) {
+                    OrderResponseDto orderResponseDto = new OrderResponseDto(order);
+                    orderResponseDtoList.add(orderResponseDto);
+                }
+            }
+            return ResponseEntity.status(HttpStatus.OK).body(orderResponseDtoList);
+        }
+    }
+
+    public ResponseEntity<List<OrderResponseDto>> getReadyOrderList() {
         List<Order> orderList = orderRepository.findAll();
         if (orderList.isEmpty()) {
             throw new NullPointerException("주문이 없습니다.");
         } else {
             List<OrderResponseDto> orderResponseDtoList = new ArrayList<>();
             for (Order order : orderList) {
-                OrderResponseDto orderResponseDto = new OrderResponseDto(order);
-                orderResponseDtoList.add(orderResponseDto);
+                if (!order.getOrderComplete()) {
+                    OrderResponseDto orderResponseDto = new OrderResponseDto(order);
+                    orderResponseDtoList.add(orderResponseDto);
+                }
             }
             return ResponseEntity.status(HttpStatus.OK).body(orderResponseDtoList);
         }
     }
 
     public ResponseEntity<List<OrderResponseDto>> getMyOrders(User user) {
-        List<Order> orderList = orderRepository.findAllByUserId(user.getId());
+        List<Order> orderList = orderRepository.findAllByUserIdOrderByCreatedAtDesc(user.getId());
         if (orderList.isEmpty()) {
             throw new NullPointerException("주문이 없습니다.");
         } else {
@@ -159,10 +185,14 @@ public class OrderService {
     public ResponseEntity<ApiResponseDto> cancelOrder(Long orderNo, User user) {
         Order order = orderRepository.findById(orderNo).orElseThrow(() -> new NullPointerException("존재하지 않는 주문 번호입니다."));
         if (user.getId().equals(order.getUser().getId()) || user.getRole().equals(UserRoleEnum.ADMIN)) {
-            orderRepository.delete(order);
-            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("주문이 취소 되었습니다.", HttpStatus.OK.value()));
+            if (order.getOrderComplete()) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ApiResponseDto("이미 완료된 주문입니다.", HttpStatus.BAD_REQUEST.value()));
+            } else {
+                orderRepository.delete(order);
+                return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("주문이 취소 되었습니다.", HttpStatus.OK.value()));
+            }
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_GATEWAY).body(new ApiResponseDto("해당 주문에 대한 권한이 없습니다.", HttpStatus.BAD_GATEWAY.value()));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ApiResponseDto("해당 주문에 대한 권한이 없습니다.", HttpStatus.BAD_REQUEST.value()));
         }
     }
 
@@ -177,7 +207,20 @@ public class OrderService {
                         tempOrderCount = tempOrderCount + orderItem.getQuantity();
                         orderItem.getProduct().setOrderCount(tempOrderCount);
                     }
-                    orderRepository.delete(order);
+                    Long tempOrderCount = order.getUser().getOrderCount();
+                    order.getUser().setOrderCount(tempOrderCount + 1);
+                    order.setOrderComplete(true);
+
+                    List<Order> totalOrderList = orderRepository.findAllByOrderByCreatedAtAsc();
+                    List<Order> completedOrderList = new ArrayList<>();
+                    for (Order order2 : totalOrderList) {
+                        if (order2.getOrderComplete()) {
+                            completedOrderList.add(order2);
+                            if (completedOrderList.size() > 100) {
+                                orderRepository.delete(completedOrderList.remove(0));
+                            }
+                        }
+                    }
                 }
             }
         }
